hero/semantic_search_faiss/requirements.txt
streamlit
pandas
numpy
faiss-cpu
sentence-transformers
openai
requests
